[logiclayer] refactors TsQuery outside of tryinto
